cumulus/pyproject.toml
[project]
name = "cumulus"
description = "CLI utility for the Nubosa system"

version = "1.0.0"
requires-python = ">=3.12.3"

dependencies = [
  "platformdirs>=4.2.0",
  "colorama>=0.4.6",
  "jinja2>=3.1.4",
  "pyyaml>=6.0.1",
]

authors = [
  {name = 'Jorge "Kutu" Dobón Blanco', email="code@dobon.dev"}
]

[project.scripts]
cumulus = 'cumulus.__main__:app'

[project.optional-dependencies]
dev = [
    "pip-tools>=7.4.1",
    "ruff>=0.1.2",
    "mypy>=1.4.1",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
cumulus = ["templates/*.jinja"]

[tool.ruff.lint]
select = ["E", "F", "I001"]

[tool.mypy]
strict = true
